Move the bulk of checkout code to makeCheckoutTransaction
